[#54676118] Change path and exec method for Puppet
Replace Vagrant's builtin provisioner for Puppet with our own shared folder
and shell script.

This will allow us to use a consistent filesystem path in Vagrant and
production. Which is useful for hiera, in particular, because unlike
extlookup the `datadir` path can't be computed relative to another Puppet
variable.

I foresee the production deployment being a pyFabric task that does a `git
checkout-index` of this repo into the same path, with any other `hierdata/`
files copied over the top, then the same script(s) run.