Give all accounts sudo
Bump to latest puppet-gds_accounts which allows multiple groups, and add all our
users to the sudo group as well as the gds group.

We're taking this approach rather than changing the sudoers file because we're
not happy with the module currently used for that in ci-puppet. We've raised a
story to look at how best to change the sudoers file instead:
https://www.pivotaltracker.com/story/show/55378608