[#54676640] Generate SSL keys pass to nginx
There is a new parameter for the mirror_environment resource
'ssl_config'. This is a hash the maps vhost name to the contents
of the key and cert.

This map is then mapped over the nginx_ssl resource which ensures
the key and cert are present. All of the files generated will end
up in /etc/nginx/ssl/, which is purged at the beginning of each
run. Both the key and the cert will have the vhost name as their name.

In order to get this working we have to use our branch of jfryman's
nginx module (https://github.com/alphagov/puppet-nginx-jfryman)
 as we needed to be able to control how the ssl files are generated
 rather than ceeding to it.

Keys are given a mode of 0400, certs 0444, /etc/nginx/ssl/ 0700 and
all are owned by root.

We have generated a development self signed key and cert that are
included in env.development.yaml