[#55707982] default vhost nginx for ssl
installing the ssl-cert which provides dummy ssl certificates for
nginx.

Using manifest to add vhost instead of hieradata. This is because the
certificates are created by ssl-cert pkg. So it is required to install
ssl-cert package first and then creating cert/key files resources and
creating nginx vhost. We are maintaining this order in manifest file.

Also nginx-puppet module has bug when we use ssl and listen options
together. We fixed this by putting a space infront of default vhost :-O