Allow vagrant user to preserve the environment
Vagrant derives it's privileges by membership of the %sudo group.
However the defaults for sudo don't include env_reset (which executes
commands in a minimal shell environment by default) or setenv (which
allows users to override env_reset if neccessary), which causes
reprovisioning a Vagrant machine to fail.

We should stop doing that.