Re-compile VirtualBox GA against current kernel
This commit recompiles VirtualBox Guest Additions against the current
kernel. It does this by installing the latest and greatest headers for the
currently-installed kernel (by querying the output of `uname -r` and using
that with Aptitude).

It also installs Dynamic Kernel Module Support which allows kernel modules
to be re-built against DKMS' separate kernel tree, thus every time the
kernel is updated, modules stored in the DKMS kernel tree are updated. This
keeps things in sync.

Finally, it forces VirtualBox to recompile itself against the DKMS kernel
tree.