Update canary response
Making this more specific means that it's quicker for us to diagnose
where this response is coming from compared to if it's just "OK".

We were seeing this response recently and nobody knew immediately that
it was coming from the mirrors. We'd have a much better chance of figuring
that out if this was the response.