Add nginx location blocks for errors
The existing errors from `/error/503.html` don't work because the
error directory isn't inside the `www_root`, it's one level higher
up (in `/srv/mirror_data`).

This commit adds location blocks for the 2 vhosts so that requests
to `/error/503.html` are served from the root directory. This means
that the file that ends up being served is `/srv/mirror_data/error/503.html`.